rename dictfile to metafile
